refactor(contacts): organize add address step
