+++ .changeset/tidy-address-flow.md
@@ -1,0 +1,5 @@
+---
+"@features/flow-contacts": patch
+---
+
+Organize Add Address flow files by responsibility
