+++ features/flow/contacts/src/steps/AddAddress/AddressEntry/Entry/ContactsAddAddressEntry.native.tsx
@@ -1,6 +1,6 @@
 import React from "react";
-import { ContactsAddAddressEntryView } from "./ContactsAddAddressEntryView.native";
-import type { ContactsAddAddressEntryProps } from "./ContactsAddAddressEntry.types";
+import { ContactsAddAddressEntryView } from "../ContactsAddAddressEntryView.native";
+import type { ContactsAddAddressEntryProps } from "../ContactsAddAddressEntry.types";
 import { useContactsAddAddressEntryViewModel } from "./useContactsAddAddressEntryViewModel.native";
 
 export function ContactsAddAddressEntry(props: ContactsAddAddressEntryProps): React.JSX.Element {
