+++ features/flow/contacts/src/steps/AddAddress/AddressEntry/Entry/ContactsAddAddressEntry.web.tsx
@@ -1,6 +1,6 @@
 import React from "react";
-import type { ContactsAddAddressEntryWebProps } from "./ContactsAddAddressEntry.web.types";
-import { ContactsAddAddressEntryView } from "./ContactsAddAddressEntryView.web";
+import type { ContactsAddAddressEntryWebProps } from "../ContactsAddAddressEntry.web.types";
+import { ContactsAddAddressEntryView } from "../ContactsAddAddressEntryView.web";
 import { useContactsAddAddressEntryViewModel } from "./useContactsAddAddressEntryViewModel.web";
 
 export function ContactsAddAddressEntry(props: ContactsAddAddressEntryWebProps): React.JSX.Element {
