+++ features/flow/contacts/src/steps/AddAddress/AddressEntry/Entry/useContactsAddAddressEntryViewModel.web.ts
@@ -3,7 +3,7 @@ import type {
   AddressLabelConfiguration,
   ContactsAddAddressEntryWebProps,
   ContactsAddAddressEntryWebViewProps,
-} from "./ContactsAddAddressEntry.web.types";
+} from "../ContactsAddAddressEntry.web.types";
 
 function getAddressLabelConfiguration({
   addressLabel,
