+++ features/flow/contacts/src/steps/AddAddress/native.ts
@@ -1,5 +1,5 @@
-export { ContactsAddAddressEntry } from "./ContactsAddAddressEntry.native";
-export type { ContactsAddAddressEntryProps } from "./ContactsAddAddressEntry.types";
+export { ContactsAddAddressEntry } from "./AddressEntry/Entry/ContactsAddAddressEntry.native";
+export type { ContactsAddAddressEntryProps } from "./AddressEntry/ContactsAddAddressEntry.types";
 export { ContactsAddAddressName } from "./AddressName/ContactsAddAddressName.native";
 export type { ContactsAddAddressNameNativeProps as ContactsAddAddressNameProps } from "./AddressName/types";
-export { ContactsAddAddressPlaceholderView } from "./ContactsAddAddressPlaceholderView.native";
+export { ContactsAddAddressPlaceholderView } from "./AddressEntry/Placeholder/ContactsAddAddressPlaceholderView.native";
