+++ features/flow/contacts/src/steps/AddAddress/web.ts
@@ -1,6 +1,6 @@
-export { ContactsAddAddressEntry } from "./ContactsAddAddressEntry.web";
-export type { ContactsAddAddressEntryWebProps as ContactsAddAddressEntryProps } from "./ContactsAddAddressEntry.web.types";
-export type { AddAddressEntryLabels as ContactsAddAddressEntryLabels } from "./types";
+export { ContactsAddAddressEntry } from "./AddressEntry/Entry/ContactsAddAddressEntry.web";
+export type { ContactsAddAddressEntryWebProps as ContactsAddAddressEntryProps } from "./AddressEntry/ContactsAddAddressEntry.web.types";
+export type { AddAddressEntryLabels as ContactsAddAddressEntryLabels } from "./Flow/types";
 export { ContactsAddAddressNameInput as ContactsAddAddressName } from "./AddressName/Input/ContactsAddAddressNameInput.web";
 export { ContactsAddAddressCompletion } from "./Completion/ContactsAddAddressCompletion.web";
 export type {
